Falta importacao

+++ src/routes.js
@@ -1,4 +1,5 @@
 import { Database } from './database.js'
+import { json } from './middlewares/json.js';
 import { buildRoutePath } from './utils/build-route-path.js';
 
 const database = new Database();
@@ -10,6 +11,10 @@ export const routes =
             path: buildRoutePath('/tasks'),
             handler: (req, res) => {
                 const { title, description } = req.body;
+
+                if (!title || !description)
+                    return res.writeHead(400).end();
+
                 database.insert({
                     title,
                     description
@@ -28,9 +33,19 @@ export const routes =
         },
         {
             method: 'PUT',
-            path: buildRoutePath('/tasks'),
+            path: buildRoutePath('/tasks/:id'),
             handler: (req, res) => {
-                const { id, title, description } = req.body;
+                const { title, description } = req.body;
+                const { id } = req.params;
+
+                if (!id || !title || !description)
+                    return res.writeHead(400).end();
+
+                const exists = database.select(id);
+
+                if (!exists)
+                    return res.end(JSON.stringify({ message: "Task not found! " }))
+
                 database.update(id, title, description);
                 return res.writeHead(200).end();
             }
@@ -40,15 +55,26 @@ export const routes =
             path: buildRoutePath('/tasks/:id'),
             handler: (req, res) => {
                 const { id } = req.params;
+                const exists = database.select(id);
+
+                if (!exists)
+                    return res.end(JSON.stringify({ message: "Task not found! " }))
+
                 database.delete(id);
                 return res.end();
             }
         },
         {
             method: 'PATCH',
-            path: buildRoutePath('/tasks/:id'),
+            path: buildRoutePath('/tasks/:id/complete'),
             handler: (req, res) => {
                 const { id } = req.params;
+
+                const exists = database.select(id);
+
+                if (!exists)
+                    return res.end(JSON.stringify({ message: "Task not found! " }))
+
                 database.completed(id);
                 return res.end();
             }
